backend/requirements.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.7.2
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
zipp==3.4.1
